# libs/openapi/swagger.yaml
openapi: 3.1.0
info:
  version: 1.0.0
  title: KVBinder API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    This API allows you to GET and PUT JSON objects based on a key.
servers:
  - url: '{server}/api'
    variables:
      server:
        default: http://localhost:3000
        description: Your server base url
tags:
  - name: Item
    description: operations on items.
paths:
  /items:
    get:
      tags:
        - Item
      summary: Get items
      description: |
        This operation retrieves all the item
      operationId: getItems
      responses:
        '200':
          description: Successfull operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
        '403':
          description: Forbidden
        '404':
          description: Item not found
      security: []
    post:
      tags:
        - Item
      summary: import a set of items
      description: |
        This operation adds all the items
      operationId: importItems
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Item'
      responses:
        '200':
          description: Successfull operation
      security: []
  /items/{key}:
    get:
      tags:
        - Item
      summary: Get item by key
      description: |
        This operation retrieves the item by its unique identifier.
      operationId: getItem
      parameters:
        - name: key
          in: path
          description: The key of the item to retrieve.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfull operation
          content:
            application/json:
              schema:
                type: object
        '403':
          description: Forbidden
        '404':
          description: Item not found
      security: []
    put:
      tags:
        - Item
      summary: Update or create an item by key
      description: |
        This operation updates or creates an item by its unique identifier.
      operationId: updateItem
      parameters:
        - name: key
          in: path
          description: The ID of the item to update.
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
        description: Updated item object
        required: true
      responses:
        '200':
          description: Successfully updated the item
        '400':
          description: Invalid key supplied
      security: []
    delete:
      tags:
        - Item
      summary: Delete an item by key
      description: This operation deletes all data associated with the requested `key`.
      operationId: deleteItem
      parameters:
        - name: key
          in: path
          description: The key of the item to delete.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully deleted a user
        '400':
          description: Invalid username provided
        '404':
          description: User not found
      security: []
  /keys:
    get:
      tags:
        - Item
      summary: Get only keys
      description: |
        This operation retrieves all the keys
      operationId: getKeys
      responses:
        '200':
          description: Successfull operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
      security: []
components:
  schemas:
    Item:
      type: object
      properties:
        key:
          type: string
        value:
          type: object
      required:
        - key
        - value
x-tagGroups:
  - name: General
    tags:
      - Item